FFC-148165: rename files
